Enable SNI in wolfSSL and use it for over the air updates
SNI is a must have these days to connect to stuff on the internet as
more and more things require it. This adds SNI handling to `ota.c` to
make sure it sends the SNI TLS extension on connecting to a secure
domain.

The new GitHub releases CDN domain requires SNI and this helps resolve
the issues with that.